fixed logging for all of storeitemcontroller

+++ TeamZealzamorpheoftheHoliestOrder-LordsoftheWesternTerritories/Controllers/StoreItemController.cs
@@ -39,30 +39,32 @@ namespace TeamZealzamorpheoftheHoliestOrder_LordsoftheWesternTerritories.Control
         [HttpGet("[action]")]
         public async Task<Category> GetCategoryById(int id)
         {
-            return await dataService.Categories.Where(i => i.Id == id).FirstOrDefaultAsync();
+            var logged = await TryLogMessage("IDataService", "Info", id.ToString(), "GetCategoryById");
+            if (logged)
+            {
+                return await dataService.Categories.Where(i => i.Id == id).FirstOrDefaultAsync();
+            }
+            else
+            {
+                return null;
+            }
         }
 
         [HttpPost("[action]")]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> AddCategory(Category category)
         {
-            if (validateClass.ValidateCategory(category))
+            var logged = await TryLogMessage("IDataService", "Info", category.Title, "AddCategory");
+            if (logged)
             {
-                await dataService.CreateCategory(category);
-                return Ok();
-            } else
-            {
-                return BadRequest();
-            }
-        }
-        [HttpPost("[action]")]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> AddItem(StoreItem item)
-        {
-            if (validateClass.ValidateStoreItem(item))
-            {
-                await dataService.CreateItem(item);
-                return Ok();
+                if (validateClass.ValidateCategory(category))
+                {
+                    await dataService.CreateCategory(category);
+                    return Ok();
+                } else
+                {
+                    return BadRequest();
+                }
             }
             else
             {
@@ -71,14 +73,46 @@ namespace TeamZealzamorpheoftheHoliestOrder_LordsoftheWesternTerritories.Control
         }
         [HttpPost("[action]")]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> AddItem(StoreItem item)
+        {
+            var logged = await TryLogMessage("IDataService", "Info", item.ItemName, "AddItem");
+            if (logged)
+            {
+                if (validateClass.ValidateStoreItem(item))
+                {
+                    await dataService.CreateItem(item);
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest();
+                }
+            }
+            else
+            {
+                return BadRequest();
+            }
+        }
+
+
+        [HttpPost("[action]")]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> ConstructACroc(CrocAttributes crocAttributes)
         {
             StoreItem item = new StoreItem($"{crocAttributes.Hobby} Croc", (decimal)49.99);
             item.Description = new DescriptionBuilderService(crocAttributes.Color, crocAttributes.Hobby).WithFancyTail(crocAttributes.Tail).WithHat(crocAttributes.Hat).WithHeldItem(crocAttributes.HeldItem).Build();
-            if (validateClass.ValidateStoreItem(item))
+            var logged = await TryLogMessage("IDataService", "Info", item.ItemName, "ConstructACroc");
+            if (logged)
             {
-                await dataService.CreateItem(item);
-                return Ok();
+                if (validateClass.ValidateStoreItem(item))
+                {
+                    await dataService.CreateItem(item);
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest();
+                }
             }
             else
             {
@@ -104,10 +138,18 @@ namespace TeamZealzamorpheoftheHoliestOrder_LordsoftheWesternTerritories.Control
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> UpdateCategory(Category category)
         {
-            if (validateClass.ValidateCategory(category))
+            var logged = await TryLogMessage("IDataService", "Info", category.Title, "UpdateCategory");
+            if (logged)
             {
-                await dataService.UpdateCategory(category);
-                return Ok();
+                if (validateClass.ValidateCategory(category))
+                {
+                    await dataService.UpdateCategory(category);
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest();
+                }
             }
             else
             {
@@ -117,13 +159,25 @@ namespace TeamZealzamorpheoftheHoliestOrder_LordsoftheWesternTerritories.Control
         [HttpPost("[action]")]
         public async Task DeleteItem(StoreItem item)
         {
-            await dataService.DeleteItem(item);
+            var logged = await TryLogMessage("IDataService", "Info", item.ItemName, "DeleteItem");
+            if (logged)
+            {
+                await dataService.DeleteItem(item);
+            }
         }
 
         [HttpPost("[action]")]
         public async Task DeleteCategory(Category category)
         {
-            await dataService.DeleteCategory(category);
+            var logged = await TryLogMessage("IDataService", "Info", category.Title, "DeleteCategory");
+            if (logged)
+            {
+                await dataService.DeleteCategory(category);
+            }
+        }
+        private async Task<bool> TryLogMessage(string service, string level, string parameters, string action)
+        {
+            return await dataService.LogMessage(new LogMessage { Service = service, Action = action, Parameters = parameters, Level = level, TimeStamp = DateTime.Now });
         }
     }
 }
